refactor: restructure game handling with new state management and handler classes
- Introduced a centralized `GameStateManager` to manage game rooms, socket mappings, and matchmaking queue.
- Created dedicated handler classes for matchmaking, game room management, moves, timers, game endings, invites, chat, and spectators to improve code organization and maintainability.
- Updated `ChessGateway` to utilize the new handlers and state manager, streamlining the connection and game logic.
- Enhanced type definitions for better clarity and type safety across the game module.
