+++ apps/api/src/game/handlers/index.ts
@@ -1,0 +1,8 @@
+export * from './matchmaking.handler';
+export * from './game-room.handler';
+export * from './move.handler';
+export * from './timer.handler';
+export * from './game-end.handler';
+export * from './invite.handler';
+export * from './chat.handler';
+export * from './spectator.handler';
