+++ apps/api/src/game/state/index.ts
@@ -1,0 +1,2 @@
+export * from './game-state.manager';
+
