+++ apps/api/src/game/types/index.ts
@@ -1,0 +1,2 @@
+export * from './game.types';
+
